+++ Tests/WalletConnectKMSTests/MockedCodec.swift
@@ -2,22 +2,22 @@
 
 import Foundation
 @testable import WalletConnectKMS
-//
+
 //class MockedCodec: Codec {
 //    var hmacAuthenticator: HMACAuthenticating
-//
+//    
 //    var encryptionPayload: EncryptionPayload!
 //    var decodedJson: String!
-//
+//    
 //    init(hmacAuthenticator: HMACAuthenticating = HMACAuthenticator()) {
 //        self.hmacAuthenticator = hmacAuthenticator
 //    }
 //
-//    func encode(plainText: String, agreementKeys: AgreementSecret) throws -> EncryptionPayload {
+//    func encode(plainText: String, symmetricKey: Data) throws -> EncryptionPayload {
 //        return encryptionPayload
 //    }
-//
-//    func decode(payload: EncryptionPayload, sharedSecret: Data) throws -> String {
+//    
+//    func decode(payload: EncryptionPayload, symmetricKey: Data) throws -> String {
 //        return decodedJson
 //    }
 //}
